pyproject.toml
[tool.poetry]
name = "microsoft-graph-webhook"
version = "0.1.0"
description = "Microsoft Graph webhook receiver for email notifications"
authors = ["Project Author <author@company.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
httpx = "^0.25.2"
python-dotenv = "^1.0.0"
pydantic = "^2.5.2"
pydantic-settings = "^2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
ruff = "^0.1.7"
mypy = "^1.7.1"
ipython = "^8.18.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["E", "F", "N", "W", "I", "UP"]
ignore = ["N803", "N815"]  # Ignore camelCase naming for API compatibility
target-version = "py312"
line-length = 88

[tool.ruff.per-file-ignores]
"src/schemas/*.py" = ["N815"]  # Allow camelCase in schema files for API compatibility
"src/routers/notifications.py" = ["N803"]  # Allow validationToken parameter